feat: setup an authentication to route PATCH users/me (earlier users/:id) and DELETE users/me (earlier users/:id)
description: override toJSON for userSchema so we can restrict to not send the valuable info of user which should not expose publically reason for this was when response is send toJSON will be called implicitly with JSON.stringify(). reason for changing the routes path is we never need user to delete/update other user info if they know the id so changed to /me so only they can only delete/update their own profile.